requirements.txt
pandas~=1.3.4
psycopg2~=2.9.2
numpy~=1.21.4
python-dotenv~=0.19.2